+++ changelog.md
@@ -4,7 +4,7 @@
 
 1. Add CharCnn
 2. Fix typo `unk_indx` to `unk_idx` of NlpContext
-3. Finish model module of `few_shot`
+3. Finish model module of `few_shot`, `style_transfer`
 
 ## 0.1.0